client: fetch client time zone on pre_connect
The current code is fetching the client time zone info at the serialization library which is incorrect and breaks the design of libfreerdp-core.
Since this timezone data is known before the connection initiates, i find it reasonable to move it to the `_preconnect` functions of clients.
